requirements.txt
Django==1.11
djangorestframework==3.9.4